==> 02_aggregate_data.R <==
# Aggregate Data

#plan(multisession, workers = parallel::detectCores() - 1)
#plan(multisession, workers = 10)

ignore_big <- F

gf_sf <- readRDS(file.path(gasflare_dir, "finaldata", "gas_flare_locations.Rds"))

### Big countries
if(ignore_big){
  adm_sf <- read_sf(file.path(wb_bound_dir, 
                              paste0("World Bank Official Boundaries - Admin ",0,".gpkg")))
  
  sf_use_s2(F)
  adm_sf$area <- adm_sf %>% st_area() %>% as.numeric()
  sf_use_s2(T)
  
  big_countries <- adm_sf %>%
    st_drop_geometry() %>%
    arrange(-area) %>%
    pull(ISO_A3) %>%
    head(15)
}
###

# Loop: ADM Level --------------------------------------------------------------
for(adm_level_i in 2){
  
  adm_sf <- read_sf(file.path(wb_bound_dir, 
                              paste0("World Bank Official Boundaries - Admin ",adm_level_i,".gpkg")))
  
  # Loop: Product --------------------------------------------------------------
  for(product_id_i in rev(c("VNP46A3", "VNP46A4"))){ 
    
    iso_all <- file.path(raster_ntl_root_dir, product_id_i) %>% 
      list.files()
  
    if(ignore_big){
      iso_all <- iso_all[!(iso_all %in% big_countries)]
    }
    
    # Loop: ISO ----------------------------------------------------------------
    for(iso_i in rev(iso_all)){

      # Ensure directories created - - - - - - - - - - - - - - - - - - - - - - -
      adm_level_name_i <- paste0("ADM", adm_level_i)
      dir.create(file.path(agg_date_dir,  adm_level_name_i), showWarnings = F)
      dir.create(file.path(agg_date_dir,  adm_level_name_i, product_id_i), showWarnings = F)
      dir.create(file.path(agg_date_dir,  adm_level_name_i, product_id_i, iso_i), showWarnings = F)
      
      # Restrict to files that don't exist - - - - - - - - - - - - - - - - - - -
      #### Raster files
      rasters_all <- file.path(raster_ntl_root_dir, product_id_i, iso_i) %>%
        list.files() 
      
      #### Prep date (from rasters)
      if(product_id_i == "VNP46A3"){
        dates_all <- str_extract(rasters_all, "\\d{4}_\\d{2}.tif") %>%
          str_replace_all(".tif", "") %>%
          str_replace_all("_", "-") %>%
          paste0("-01")
      }
      
      if(product_id_i == "VNP46A4"){
        dates_all <- str_extract(rasters_all, "\\d{4}.tif") %>%
          str_replace_all(".tif", "") %>%
          as.numeric()
      }
      
      #### Out paths
      out_paths <- file.path(agg_date_dir, adm_level_name_i, product_id_i, iso_i, paste0(dates_all, ".Rds"))
      
      #### Restrict to files that don't already exist
      files_not_processed_tf <- !file.exists(out_paths)
      
      rasters_not_processed <- rasters_all[files_not_processed_tf]
      
      if(length(rasters_not_processed) > 0){
        
        roi_sf <- adm_sf[adm_sf$ISO_A3 %in% iso_i,] %>%
          st_make_valid()
        
        # Loop: Date -------------------------------------------------------------
        for(raster_i in rasters_not_processed){
          message(paste0("Processing: ", iso_i))
          #future_map(rasters_not_processed, function(raster_i) {
          
          tmp <- tryCatch(
            {
              
              #### Prep out file
              ## Date from single raster
              if(product_id_i == "VNP46A3"){
                date_i <- str_extract(raster_i, "\\d{4}_\\d{2}.tif") %>%
                  str_replace_all(".tif", "") %>%
                  str_replace_all("_", "-") %>%
                  paste0("-01")
              }
              
              if(product_id_i == "VNP46A4"){
                date_i <- str_extract(raster_i, "\\d{4}.tif") %>%
                  str_replace_all(".tif", "") %>%
                  as.numeric()
              }
              
              # Prep out file - - - - - - - - - - - - - - - - - - - - - - - - - - - - - 
              OUT_PATH <- file.path(agg_date_dir, adm_level_name_i, product_id_i, iso_i, paste0(date_i, ".Rds"))
              
              message(paste0("Processing: ", adm_level_name_i, " - ", product_id_i, " - ", iso_i, " - ", date_i))
              
              # Prep raster - - - - - - - - - - - - - - - - - - - - - - - - - - - - - 
              #### NTL
              r <- rast(file.path(raster_ntl_root_dir, product_id_i, iso_i, raster_i))
              
              #### Quality
              raster_qual_i <- raster_i %>%
                str_replace_all("Snow_Free", "Snow_Free_Quality")
              raster_qual_i_path <- file.path(raster_qual_root_dir, product_id_i, iso_i, raster_qual_i)
              
              #if(!file.exists(raster_qual_i_path)) next
              
              qual_r <- rast(raster_qual_i_path)
              
              #### Crop/Mask
              r      <- r      %>% crop(roi_sf) %>% mask(roi_sf)
              qual_r <- qual_r %>% crop(roi_sf) %>% mask(roi_sf)
              
              #### Prep Gas Flaring
              # There were some issues of "Loop X is not valid"; 
              # Doing: %>% st_combine() %>% st_make_valid() resolves
              inter_tf <- tryCatch(
                {
                  # First attempt
                  st_intersects(roi_sf, gf_sf, sparse = FALSE) %>%
                    apply(2, sum) %>%
                    as.vector()
                },
                error = function(e) {
                  #message("Primary st_intersects() failed — using combined/made-valid/unioned roi_sf")
                  
                  st_intersects(
                    roi_sf %>% st_combine() %>% st_make_valid(),
                    gf_sf,
                    sparse = FALSE
                  ) %>%
                    apply(2, sum) %>%
                    as.vector()
                }
              )
              
              gf_sf_i <- gf_sf[inter_tf > 0,]
              
              rm(inter_tf)
              
              # Extract - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
              METRICS <- c("sum", "mean", "median", "max", "quantile")
              QUANTILES <- c(0.05, 0.95)
              roi_df <- roi_sf %>% st_drop_geometry()
              
              ntl_df <- exact_extract(r, roi_sf, METRICS, quantiles = QUANTILES, progress = F) %>%
                rename_with(~ paste0("ntl_", .x)) 
              
              # Extract Gas Flaring - - - - - - - - - - - - - - - - - - - - - - - - 
              if(nrow(gf_sf_i) > 0){
                
                ## Gas Flaring
                gf_sf_5km_i  <- gf_sf_i %>% st_buffer(dist = 5000)
                gf_sf_10km_i <- gf_sf_i %>% st_buffer(dist = 10000)
                
                ## Rasters
                r_gf_5km   <- r %>% mask(gf_sf_5km_i)
                r_nogf_5km <- r %>% mask(gf_sf_5km_i, inverse = T)
                
                r_gf_10km   <- r %>% mask(gf_sf_10km_i)
                r_nogf_10km <- r %>% mask(gf_sf_10km_i, inverse = T)
                
                ## Extract
                ntl_gf_5km_df <- exact_extract(r_gf_5km, roi_sf, METRICS, quantiles = QUANTILES, progress = F) %>%
                  rename_with(~ paste0("ntl_gf_5km_", .x)) 
                
                ntl_nogf_5km_df <- exact_extract(r_nogf_5km, roi_sf, METRICS, quantiles = QUANTILES, progress = F) %>%
                  rename_with(~ paste0("ntl_nogf_5km_", .x)) 
                
                ntl_gf_10km_df <- exact_extract(r_gf_10km, roi_sf, METRICS, quantiles = QUANTILES, progress = F) %>%
                  rename_with(~ paste0("ntl_gf_10km_", .x)) 
                
                ntl_nogf_10km_df <- exact_extract(r_nogf_10km, roi_sf, METRICS, quantiles = QUANTILES, progress = F) %>%
                  rename_with(~ paste0("ntl_nogf_10km_", .x)) 
                
              } else{
                
                #### No Gas Flares [Same as total NTL]
                ntl_nogf_5km_df  <- ntl_df
                ntl_nogf_10km_df <- ntl_df
                
                names(ntl_nogf_5km_df) <- names(ntl_nogf_5km_df) %>%
                  str_replace_all("ntl_", "ntl_nogf_5km_")
                
                names(ntl_nogf_10km_df) <- names(ntl_nogf_10km_df) %>%
                  str_replace_all("ntl_", "ntl_nogf_10km_")
                
                #### With Gas Flares
                ntl_gf_5km_df  <- ntl_df
                ntl_gf_10km_df <- ntl_df
                
                names(ntl_gf_5km_df) <- names(ntl_gf_5km_df) %>%
                  str_replace_all("ntl_", "ntl_gf_5km_")
                
                names(ntl_gf_10km_df) <- names(ntl_gf_10km_df) %>%
                  str_replace_all("ntl_", "ntl_gf_10km_")
                
                ntl_gf_5km_df[]  <- NA
                ntl_gf_10km_df[] <- NA
                
              }
              
              # Extract Quality - - - - - - - - - - - - - - - - - - - - - - - - - -
              ntl_quality_df <- exact_extract(qual_r, 
                                              roi_sf, 
                                              function(values, coverage_fraction) {
                                                # ntl_quality_prop_0_good <- sum( (values %in% 0) * coverage_fraction, na.rm = TRUE) /
                                                #   sum( !is.na(values) * coverage_fraction, na.rm = TRUE)
                                                # 
                                                # ntl_quality_prop_1_poor <- sum( (values %in% 1) * coverage_fraction, na.rm = TRUE) /
                                                #   sum( !is.na(values) * coverage_fraction, na.rm = TRUE)
                                                # 
                                                # ntl_quality_prop_2_gapfilled <- sum( (values %in% 2) * coverage_fraction, na.rm = TRUE) /
                                                #   sum( !is.na(values) * coverage_fraction, na.rm = TRUE)
                                                
                                                ntl_quality_prop_0_good <- sum( (values %in% 0) , na.rm = TRUE) /
                                                  sum( length(values) , na.rm = TRUE)
                                                
                                                ntl_quality_prop_1_poor <- sum( (values %in% 1) , na.rm = TRUE) /
                                                  sum( length(values) , na.rm = TRUE)
                                                
                                                ntl_quality_prop_2_gapfilled <- sum( (values %in% 2) , na.rm = TRUE) /
                                                  sum( length(values) , na.rm = TRUE)
                                                
                                                data.frame(ntl_quality_prop_0_good, 
                                                           ntl_quality_prop_1_poor, 
                                                           ntl_quality_prop_2_gapfilled)
                                              },
                                              progress = F)
              
              # Extract Proportion NA - - - - - - - - - - - - - - - - - - - - - - - 
              ntl_prop_na_df <- exact_extract(r, 
                                              roi_sf, 
                                              function(values, coverage_fraction) {
                                                # ntl_quality_prop_na <- sum( is.na(values) * coverage_fraction, na.rm = TRUE) /
                                                #   sum( !is.na(values) * coverage_fraction, na.rm = TRUE)
                                                
                                                ntl_quality_prop_na <- sum( is.na(values), na.rm = TRUE) /
                                                  sum( length(values) , na.rm = TRUE)
                                                
                                                data.frame(ntl_quality_prop_na)
                                              },
                                              progress = F)
              
              # Prep output - - - - - - - - - - - - - - - - - - - - - - - - - - - - 
              if(product_id_i == "VNP46A3"){
                date_i <- date_i %>% ymd()
              }
              
              roi_df$date <- date_i
              roi_df$n_gasflaring_locs <- nrow(gf_sf_i)
              
              roi_df <- bind_cols(roi_df, 
                                  ntl_prop_na_df,
                                  ntl_quality_df,
                                  ntl_df,
                                  ntl_gf_5km_df,
                                  ntl_nogf_5km_df,
                                  ntl_gf_10km_df,
                                  ntl_nogf_10km_df)
              
              
              # Export - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - 
              saveRDS(roi_df, OUT_PATH)
              
              ## end try catch
            },
            error = function(e) {
              message("rror occurred:")
              message(conditionMessage(e))
              return(NULL)
            }
          )
          ## end try catch
          
          
          #})
        }
        
        
        
      }
    }
  }
}
#}
